test: switch runners
Remove playwright in favor of node's built in test runner.
It is simpler to maintain since we only need a small shim for
local and session storage in place. Right now that is setup
using JSDOM, but is a candidate for removal later to just
mock internally with Sinon.

Test coverage expanded to 100% baseline. In the process,
updated some of the validation logic to throw errors. Also
now handling "not before" as it came up in unit testing.